requirements.txt
# API
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# Visão computacional
ultralytics>=8.1.0
opencv-python>=4.9.0.80
Pillow>=10.3.0

# Audio
openai-whisper>=20231117
soundfile>=0.12.1
librosa>=0.10.1

# Azure
azure-cognitiveservices-speech>=1.35.0
azure-ai-textanalytics>=5.3.0
azure-ai-vision-imageanalysis>=1.0.0
openai>=1.12.0

# Relatórios
reportlab>=4.1.0

# ML
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.4
pandas>=2.2.0
